requirements.txt
datasets
torch==2.5.1
transformers==4.47.0
sentence-transformers==3.3.1
openai
pinecone-client==5.0.1
protoc-gen-openapiv2==0.0.1
cohere
grpcio
python-dotenv
pytest
locust==2.32.4